BTK_MachineLearning/Week4/Day1/requirements.txt
pandas
plotly
streamlit
yfinance
pandas-datareader
prophet
nltk
stylecloud
pillow==9.4
Image